+++ src/main.tsx
@@ -4,11 +4,14 @@ import App from './App.tsx';
 import './styles/index.scss';
 import ErrorBoundary from './components/ErrorBoundary/ErrorBoundary.tsx';
 import Error from './components/Error/Error.tsx';
+import { BrowserRouter } from 'react-router-dom';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
-    <ErrorBoundary fallback={<Error />}>
-      <App />
-    </ErrorBoundary>
+    <BrowserRouter>
+      <ErrorBoundary fallback={<Error />}>
+        <App />
+      </ErrorBoundary>
+    </BrowserRouter>
   </React.StrictMode>
 );
