+++ src/styles/index.scss
@@ -1,2 +1,3 @@
 @import './global/reset';
 @import './global/global';
+@import './global/app';
